Small change for convenience in GLO CDMA
